requirements.txt
pandas
numpy
torch
scikit-learn
langchain
langchain-google-generative-ai
python-dotenv
torchvision